Update Linear_System_Solver_LU_Crout
Modified the code so that it actually works when flashed to the Arduino. For some reason though, solving this linear system is not accurate. If A*x is performed to try to return b, what should be 2 in the first array cell is around 5.

Anyone know why this algorithm isn't working?