Update Linear_System_Solver_LU_Crout
This is the first working code of the linear system solver!
A concern addressed previously was that MATLAB and C++/Arduino did not have the same output. However I realized two things:
(1) The first was that I wasn't printing all the decimal points of the float. This made it seemingly inaccurate.
(2) MATLAB created a different solution because it does partial pivoting. The solution generated by C++/Arduino is satisfactory.

Since the code works, I timed it and we're at about 1 ms of computation time. My thought is that this is alright, but that we could possibly drive this value down by not using floats. For the moment though. the code seems satisfactory for use.