Open Loop Control strategy
The included Arduino sketch includes code to take transmitter "Mode 2" control inputs, and fuse them together to select the speeds of the four motors on the drone. This is open-loop since it directly reads the inputs, and does not actually check to see if the inputs are correct from feedback.

It would be a great help if there was a way to reduce the number of floating-point operations to get the four speeds. I think by using vectors it could be possible, but it needs to be tested and flushed out extensively.