Add files via upload
Code by Iwan Martin (1/16/17)
"Some notes: we found that the ESC didn't like analogWrite. So we used the Servo library instead to create pin outputs. The ESC's would also not accept a value that was too low so we configured it to use a range from 20-180 instead of 0-180, and that seemed to make it happy."