Add files via upload
Version 2 intends to be a complete solution to flying the drone. Signal collection is combined with the math behind the open-loop control, in addition to outputting the values to Servo. This still needs to be tested on hardware to verify if it works, and appropriate choices for some constants in the math must be done.